solved casting problems

+++ Core/Darkorbit/Commands/PostHandshake/ReviveShip.cs
@@ -31,7 +31,7 @@ namespace PolskaBot.Core.Darkorbit.Commands.PostHandshake
 
         public override void Write()
         {
-            packetWriter.Write((short)26 + Version.Length + SID.Length);
+            packetWriter.Write((short)(26 + Version.Length + SID.Length));
             packetWriter.Write(ID);
             packetWriter.Write((short)24278); //class_356.ID
             packetWriter.Write(Selected);
@@ -39,7 +39,7 @@ namespace PolskaBot.Core.Darkorbit.Commands.PostHandshake
             packetWriter.Write(Login.ID);
             packetWriter.Write((short)-19684);
             packetWriter.Write((int)((uint)InstanceID << 6 | (uint)InstanceID >> 26));
-            packetWriter.Write((short)FactionID);
+            packetWriter.Write(FactionID);
             packetWriter.Write((int)((uint)UserID >> 3 | (uint)UserID << 29));
             packetWriter.Write((short)Version.Length);
             packetWriter.Write(Encoding.UTF8.GetBytes(Version));
